requirements.txt
# ===============================
# Core
# ===============================
python-dotenv>=1.0.1
loguru>=0.7.2

# ===============================
# HTTP / API
# ===============================
requests>=2.31.0
httpx>=0.27.0

# ===============================
# Telegram (async)
# ===============================
aiogram>=3.4.1

# ===============================
# Data & Analysis
# ===============================
pandas>=2.2.0
numpy>=1.26.3
ta>=0.11.0

# ===============================
# Database (PostgreSQL)
# ===============================
SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9



# ===============================
# Utils
# ===============================
python-dateutil>=2.9.0
pytz>=2024.1
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.1
python-telegram-bot==20.7

grpcio>=1.60.0
grpcio-tools>=1.60.0
python-telegram-bot==20.7
python-dotenv>=1.2.1